optimized Navigation folder
